源码/requirement.txt
pymycobot==3.2.4
numpy
opencv-python==4.5.5.62
opencv-contrib-python==4.5.5.62
pyserial
